refactor capabilities validation
